Classes are set and UI functions as expected, ready for next integration phase
