+++ tasks.py
@@ -146,4 +146,5 @@ def style(c):
     c.run('pip install -U flake8', hide=True)
     print("Running PEP style checks...")
     c.run('flake8 tasks.py run_tests.py kintree_gui.py kintree/kintree_gui.py kintree/setup_inventree.py \
-        kintree/common/ kintree/config/ kintree/database/ kintree/kicad/*.py kintree/search/*.py')
+        kintree/common/ kintree/config/ kintree/database/ kintree/kicad/*.py kintree/search/*.py \
+        kintree/gui/gui.py kintree/gui/views.py')
